# perf: Cursor based lexer