--- a/ast/src/lib.rs
+++ b/ast/src/lib.rs
@@ -20,11 +20,6 @@ mod generic;
 mod impls;
 mod ranged;
 
-#[cfg(feature = "malachite-bigint")]
-pub use malachite_bigint as bigint;
-#[cfg(all(feature = "num-bigint", not(feature = "malachite-bigint")))]
-pub use num_bigint as bigint;
-
 pub use builtin::*;
 pub use generic::*;
 pub use ranged::Ranged;