--- a/ast/src/ranged.rs
+++ b/ast/src/ranged.rs
@@ -2,8 +2,6 @@
 
 use crate::text_size::{TextRange, TextSize};
 
-pub use crate::builtin::*;
-
 pub trait Ranged {
     fn range(&self) -> TextRange;