… .github/workflows/ci.yaml
… Cargo.toml
… ast/Cargo.toml
… core/Cargo.toml
… format/Cargo.toml
… parser/Cargo.toml
… parser/src/lexer/indentation.rs
… parser/src/python.lalrpop
… parser/src/python.rs
… parser/src/snapshots/rustpython_parser__parser__tests__numeric_literals.snap
… parser/src/snapshots/rustpython_parser__parser__tests__numeric_literals_attribute_access.snap
… parser/src/snapshots/rustpython_parser__parser__tests__parse_class_generic_types.snap
… parser/src/snapshots/rustpython_parser__parser__tests__parse_function_definition.snap
… parser/src/string.rs
… parser/src/token.rs