@MichaReiser 2023-07-23
Current dependencies on/for this PR:
* main
  * **PR #38** <a href="https://app.graphite.dev/github/pr/astral-sh/RustPython-Parser/38" target="_blank"><img src="https://static.graphite.dev/graphite-32x32.png" alt="Graphite" width="10px" height="10px"/></a>  👈

This comment was auto-generated by [Graphite](https://app.graphite.dev/github/pr/astral-sh/RustPython-Parser/38?utm_source=stack-comment).