@MichaReiser 2023-07-23
I'm opening this for review. It's not done done but I'm interested in getting some feedback.