@charliermarsh 2023-07-23
> Does the ecosystem check report new syntax errors?

It _should_ because we track a diagnostic for those (E999), and we'd expect to see some other diagnostics "disappear".