@dhruvmanila 2023-07-24
> * Use `&'str` for Comment, String, and Identifier: The lexer doesn't perform any normalization on Strings, Comments, and Identifiers.

Might be able to use `&'str` for MagicCommand as well ;)