@MichaReiser 2023-07-24
> > * Use `&'str` for Comment, String, and Identifier: The lexer doesn't perform any normalization on Strings, Comments, and Identifiers.
> 
> Might be able to use `&'str` for MagicCommand as well ;)

There's a subtle difference between how our lexer handles `MagicCommand` and `String` related to line continuation:
* Magic Command: Omits the line continuation and newline token in the AST
* String: The value string retains the line continuation and newline token

To me, the behavior of Magic commands seems more in line with how the lexer handles line continuation (it omits them). But it does mean we need to use a cow because it is sometimes necessary to drop a few characters.