@MichaReiser 2023-07-24
> I'm planning to leave the in-depth lexer review to @konstin since it looks like he's already read through it in detail, but can you tag me if there are any specific things you want input on? And/or if you feel another close review is needed?

I would be interested in your perspective on whether `String` should normalize newlines or not.